Alumnos/FS/ELOY-MARTINEZ-GOMEZ/Kafka/requirements.txt
mysql.connector
mysql-connector-python
kafka-python
confluent_kafka
requests